Fixed mistakes according to DmitriyVTitov's comments
